perform + secure + sonar

+++ modules/backend-statistic-rest/src/main/java/org/opencps/statistic/rest/engine/service/StatisticSumYearCalcular.java
@@ -2632,7 +2632,7 @@ public class StatisticSumYearCalcular {
 		
 		/* value get in the latest month */
 		if (releaseCount > 0) {
-			ontimePercentage = Math.round(((betimesCount + ontimeCount)*100)/releaseCount);
+			ontimePercentage = Math.round( (float)((betimesCount + ontimeCount)*100)/releaseCount);
 		}
 		
 		dossierStatisticData.setTotalCount(totalCount);
@@ -2788,7 +2788,7 @@ public class StatisticSumYearCalcular {
 		
 		/* value get in the latest month */
 		if (releaseCount > 0) {
-			ontimePercentage = Math.round(((betimesCount + ontimeCount)*100)/releaseCount);
+			ontimePercentage = Math.round( (float)((betimesCount + ontimeCount)*100)/releaseCount);
 		}
 		
 		dossierStatisticData.setTotalCount(totalCount);
